cli-client/requirements.txt
httpx
python-dotenv
textual